Moved plotting utilities into plotting.py
